requirements.txt
asgiref==3.5.2
coverage==6.4.4
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
